+++ packages/rich-text-from-markdown/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [16.2.0](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@16.1.8...@contentful/rich-text-from-markdown@16.2.0) (2026-04-08)
+
+### Features
+
+- add esm support [ZEND-7778] ([#1069](https://github.com/contentful/rich-text/issues/1069)) ([8c320bd](https://github.com/contentful/rich-text/commit/8c320bde7fa313572bdad84b91a6fd12224afc3a)), closes [#1068](https://github.com/contentful/rich-text/issues/1068)
+
+## [16.1.8](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@16.1.7...@contentful/rich-text-from-markdown@16.1.8) (2025-11-04)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [16.1.7](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@16.1.6...@contentful/rich-text-from-markdown@16.1.7) (2025-09-23)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [16.1.6](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@16.1.5...@contentful/rich-text-from-markdown@16.1.6) (2025-09-23)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [16.1.5](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@16.1.4...@contentful/rich-text-from-markdown@16.1.5) (2025-09-10)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [16.1.4](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@16.1.3...@contentful/rich-text-from-markdown@16.1.4) (2025-09-09)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [16.1.3](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@16.1.2...@contentful/rich-text-from-markdown@16.1.3) (2025-09-04)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [16.1.2](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@16.1.1...@contentful/rich-text-from-markdown@16.1.2) (2025-08-25)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+## [16.1.1](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@16.1.0...@contentful/rich-text-from-markdown@16.1.1) (2025-07-15)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
+# [16.1.0](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@16.0.1...@contentful/rich-text-from-markdown@16.1.0) (2025-07-02)
+
+### Features
+
+- **markdown:** transform strikethrough to rich-text mark ([#883](https://github.com/contentful/rich-text/issues/883)) ([ac51579](https://github.com/contentful/rich-text/commit/ac51579f5e716bb15862796177a924c7e838edb6))
+
+## [16.0.1](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@16.0.0...@contentful/rich-text-from-markdown@16.0.1) (2025-06-16)
+
+**Note:** Version bump only for package @contentful/rich-text-from-markdown
+
 # [16.0.0](https://github.com/contentful/rich-text/compare/@contentful/rich-text-from-markdown@15.18.10...@contentful/rich-text-from-markdown@16.0.0) (2024-10-29)
 
 ### Features